MentorshipTasks/SysAdmin/chat_server/app/requirements.txt
flask
flask-socketio
eventlet
mysql-connector-python

